Switching to Andrew Gerrard's Golang CoC to use the Go ecosystem standard
Hopefully no one has any problems with the change, but if anyone feels
that the Stumptown Syndicate CoC is superior please feel free to explain
why and it will be taken into consideration.
If anyone has issues with the change, feel free to contact me:
joshroppo@gmail.com. I'm happy to hear other points of view.